chore: move pixi file from graph to data-display
