+++ v3/src/components/data-display/hooks/use-connecting-lines.ts
@@ -2,7 +2,7 @@ import { curveLinear, line, select } from "d3"
 import { tip as d3tip } from "d3-v6-tip"
 import { useCallback } from "react"
 import { useDataConfigurationContext } from "./use-data-configuration-context"
-import { PixiBackgroundPassThroughEvent, PixiPoints } from "../../graph/utilities/pixi-points"
+import { PixiBackgroundPassThroughEvent, PixiPoints } from "../pixi/pixi-points"
 import { t } from "../../../utilities/translation/translate"
 import { IConnectingLineDescription, transitionDuration } from "../data-display-types"
