+++ v3/src/components/graph/components/casedots.tsx
@@ -11,7 +11,7 @@ import {useDataSetContext} from "../../../hooks/use-data-set-context"
 import {useGraphContentModelContext} from "../hooks/use-graph-content-model-context"
 import {useGraphLayoutContext} from "../hooks/use-graph-layout-context"
 import {setPointCoordinates} from "../utilities/graph-utils"
-import {IPixiPointMetadata, PixiPoints} from "../utilities/pixi-points"
+import {IPixiPointMetadata, PixiPoints} from "../../data-display/pixi/pixi-points"
 
 export const CaseDots = function CaseDots(props: {
   pixiPoints?: PixiPoints
