+++ v3/src/components/graph/components/freedotplotdots.tsx
@@ -12,7 +12,7 @@ import {useDataSetContext} from "../../../hooks/use-data-set-context"
 import {useGraphContentModelContext} from "../hooks/use-graph-content-model-context"
 import {useGraphLayoutContext} from "../hooks/use-graph-layout-context"
 import {setNiceDomain, setPointCoordinates} from "../utilities/graph-utils"
-import {circleAnchor, hBarAnchor, vBarAnchor} from "../utilities/pixi-points"
+import {circleAnchor, hBarAnchor, vBarAnchor} from "../../data-display/pixi/pixi-points"
 import { computeBinPlacements, computePrimaryCoord, computeSecondaryCoord } from "../utilities/dot-plot-utils"
 import { useDotPlotDragDrop } from "../hooks/use-dot-plot-drag-drop"
 import { AxisPlace } from "../../axis/axis-types"
