+++ v3/src/components/graph/graphing-types.ts
@@ -1,4 +1,4 @@
-import { PixiPoints } from "./utilities/pixi-points"
+import { PixiPoints } from "../data-display/pixi/pixi-points"
 
 export interface PlotProps {
   pixiPoints?: PixiPoints
