+++ v3/src/components/graph/hooks/use-chart-dots.ts
@@ -4,7 +4,7 @@ import { useDataSetContext } from "../../../hooks/use-data-set-context"
 import { setPointSelection } from "../../data-display/data-display-utils"
 import { useDataDisplayAnimation } from "../../data-display/hooks/use-data-display-animation"
 import { SubPlotCells } from "../models/sub-plot-cells"
-import { PixiPoints } from "../utilities/pixi-points"
+import { PixiPoints } from "../../data-display/pixi/pixi-points"
 import { useGraphContentModelContext } from "./use-graph-content-model-context"
 import { useGraphDataConfigurationContext } from "./use-graph-data-configuration-context"
 import { useGraphLayoutContext } from "./use-graph-layout-context"
