+++ v3/src/components/map/components/map-background.tsx
@@ -4,7 +4,7 @@ import { useMemo } from "use-memo-one"
 import { Background } from "../../data-display/components/background"
 import { Marquee } from "../../data-display/components/marquee"
 import { MarqueeState } from "../../data-display/models/marquee-state"
-import { IPixiPointsArrayRef } from "../../graph/utilities/pixi-points"
+import { IPixiPointsArrayRef } from "../../data-display/pixi/pixi-points"
 import { IMapContentModel } from "../models/map-content-model"
 import { mstReaction } from "../../../utilities/mst-reaction"
