+++ v3/src/components/map/components/map-polygon-layer.tsx
@@ -9,7 +9,7 @@ import {DEBUG_MAP, debugLog} from "../../../lib/debug"
 import {isSelectionAction, isSetCaseValuesAction} from "../../../models/data/data-set-actions"
 import {transitionDuration} from "../../data-display/data-display-types"
 import {handleClickOnCase} from "../../data-display/data-display-utils"
-import { PixiBackgroundPassThroughEvent } from "../../graph/utilities/pixi-points"
+import { PixiBackgroundPassThroughEvent } from "../../data-display/pixi/pixi-points"
 import {useDataDisplayLayout} from "../../data-display/hooks/use-data-display-layout"
 import {useMapModelContext} from "../hooks/use-map-model-context"
 import {IMapPolygonLayerModel} from "../models/map-polygon-layer-model"
